Pull in newer MRI code
The MRI file layout has been updated to be more compatible with the
Arduino build system. Rather than adding core/ to the INCLUDE path,
the root of the MRI project is instead added and the core/ is
explicitly prepended to the header names when #include'ing them.

I updated the copyright messages to use my github account URL instead
of the mbed URL.

I also use #include <> and #include "" more consistently than I did
previously.

Note: This commit also includes the removal of a lot of trailing
      whitespace that my old editor setup used to leave around in my
      source files.

Update MRI version to 1.1 for changes being worked on for this
Portenta project.