Add -mfpu= flag when assembling .S files
The assembly language context saving/restoring code in MRI contains a
few floating point instructions to move FPU registers in/out of the
context on Cortex-M4F and Cortex-M7F cores. For these to assemble
successfully the -mfpu flag must be set correctly when assembling .S
files. This flag is already set correctly when compiling .c and .cpp
files. This commit just adds it to the assembler command line along
with the -mcpu flag that was already there.