DebugSerial can now be global
DebugSerial can now be declared as a global and it will take care of
calling begin() on the underlying Serial object. So now if you declare
a DebugSerial object somewhere in your code, that will be enough to
pull the MRI debug monitor into your code so that you can use GDB for
remote debugging.

The implementation of calling HardwareSerial::begin() isn't what I
would describe as simple:
* At time the DebugSerial object's constructor runs, the underlying
  HardwareSerial object might not yet have been initialized. It is
  safest to perform this call from setup().
* DebugSerial sets a temporary breakpoint on setup() from its global
  constructor. When this breakpoint hits, it can then call the begin()
  method.
* The one trick is that this begin() call can't be made from within MRI
  itself since it performs a memory allocation which isn't allowed to
  happen when MRI is running in Handler Mode. To workaround this issue,
  the register context is saved and then the necessary registers
  modified in the context to simulate a call to begin(). Execution is
  resumed and the begin() call happens in thread mode as though it was
  called on the first line of setup(). When it returns back to setup(),
  a breakpoint is again hit and the context restored to its original
  state.

By default the DebugSerial object will cause execution to be halted on
the first line of setup() so that the user can connect with GDB and
set breakpoints/watchpoints at points of interest before they allow
their program to run. This can be disabled via the 4th parameter
of DebugSerial's constructor.