Initial commit of ThreadMRI
This is just the beginning of the thread mode version of MRI. I have
just stubbed out most of the Platform* functions for now so that I can
enter the main MRI loop and have GDB connect successfully. The only
real GDB commands that it supports properly at this point in time are
commands which read/write memory. Everthing else like register read/
write are just being faked.

NOTE: I wanted to start this prototype having it use the USBSerial
      object but I am hitting a major issue with its driver at this
      time so I switched to UART1 instead to unblock myself.