Fake break into main() thread
I have added code to the prototype which can suspend all threads except
for mriMain(), the thread actually used for debugging. It then reads
the context for the main thread out of RTX so that it looks like a
breakpoint was encountered in main() when GDB connects. This code
shows that it is possible to determine the state of RTX threads and
even read their context.