Support CTRL+C halting of running threads
Hook the USART ISR and pend a DebugMon interrupt if GDB sends data
while the debuggee is running, typically this will be a CTRL+C but
stopping for any command is a good thing.

Now uses the CORTEXM_FLAGS_CTRL_C flag to determine if SIGINT should
be returned from Platform_DetermineCauseOfException() as the cause of a
debug halt. Previously it was hardcoded to use the exception numbers
of the LPC176x UART handlers.