ThreadMRI can now be global
ThreadMRI can now be declared as a global like DebugSerial. It will
take care of calling begin() on the underlying Serial object. So now if
you declare a ThreadMRI object somewhere in your code, that will be
enough to pull the MRI debug monitor into your code so that you can use
GDB for remote debugging.

The implementation of calling HardwareSerial::begin() takes a bit of
extra work:
* At time the ThreadMRI object's constructor runs, the underlying
  HardwareSerial object might not yet have been initialized. It is
  safest to perform this call from setup().
* ThreadMRI sets a temporary breakpoint on setup() from its global
  constructor. When this breakpoint hits, it can then call the begin()
  method.

By default the ThreadMRI object will cause execution to be halted on
the first line of setup() so that the user can connect with GDB and
set breakpoints/watchpoints at points of interest before they allow
their program to run. This can be disabled via the 4th parameter
of ThreadMRI's constructor.

I doubled the stack size from 512 to 1024 bytes for mriMain() as the
STM code to initialize the baud rate of the serial port is pretty stack
hungry and it overflowed at 512 bytes.