Switch sample to use SerialUSB
I commented out the UART1 test and switched to using SerialUSB instead.
This work with my setup when I use a full-speed hub but not via high
speed USB. I am starting to get suspicious that it might have something
to do with my USB-C cable so I am pushing up this change for others to
try.