Thread context should always contain FP regisers
RTX only stacks the floating point registers when it detects that a
thread makes use of them. Previously ThreadMRI would only return the
FP registers in the context if the thread had them but GDB will
actually return an error if it sees the size of the context change
between debug halts. This commit will return a dummy set of zeroed out
float registers to GDB when the RTX thread doesn't have any of its own
to return.