Fix bugs in GDB mem read fault detection
While testing the lastest version of ThreadMRI I noticed that it was
hanging when I would ask GDB to read from an invalid memory address.
It turns out that advancePCToNextInstruction() was using the XPSR
register instead of the PC. I also added code to properly clear the
fault status registers when GDB memory read/write faults are detected.