Set serial irq priority from setup()
Previously the interrupt priority of the serial peripheral was only set
from Platform_Init() which is only called from the DebugSerial
constructor. When constructed as a global, it is possible that the
serial driver will overwrite the desired priority. This commit also
sets the desired priority from DebugSerial::initSerial() which will be
called later from setup() when the DebugSerial object is global.

While this doesn't happen for the UART Serial objects, I did hit it
when experimenting with USBSerial since it sets its interrupt priority
to 1 when I wanted it to be 0.