Make it easier to see which threads aren't suspended
I have created a g_threadNamesToIgnore[] array at the top of
ThreadDebug.cpp listing the names of the threads not to be suspended
when a debug event occurs. Currently the list only contains "rtx_idle".
Before the list was hardcoded in the suspendAllApplicationThreads()
function, where it was hard for maintainers to find.