Add detach and monitor support
The "detach" command allows GDB to detach from a running program linked
with MRI. It causes the debuggee to resume execution and then the user
can exit GDB with the program on the embedded device in a running state
whereas today it would remain halted when GDB disconnects.

Added support for the following "monitor" commands:
* "monitor help" - Lists the currently supported monitor commands to
  the GDB console. Issuing an unknown monitor command will also display
  this help list.
* "monitor showfault" - This command will dump information about the
  current fault to the GDB console just like what happens if already
  connected when the fault occurs. This is useful if you are attaching
  GDB after the fault happens.
* "monitor reset" - After issuing this command in GDB the
  microcontroller will reset on the next "continue" or "detach". This
  doesn't work well when using USBSerial as the debug interface since
  the reset will cause the TTY device being used by GDB to briefly
  disappear during the reset, causing GDB to disconnect. The user
  needs to reissue the appropriate "target remote /dev/tty*" to make
  the connection again.

I needed to implement mriPlatform_CommHasTransmitCompleted() API for
both ThreadDebug and KernelDebug to indicate if the last text
transmitted has been sent. This is needed to make sure that the ack
for the "continue" command is sent back to GDB before the device is
actually reset or GDB gets confused.