Cleanup thread suspension code
Previously I was using the osThreadSuspend() and osThreadResume() APIs
to freeze and thaw the application threads during debug halt. I code
reviewed the RTX implementation of osThreadResume() and noticed that it
incorrectly resumed previously waiting threads to the ready state. The
code now suspends the application threads by lowering their priority to
osPriorityIdle and then elevating the priority of the idle thread to
osPriorityIdle+1 so the idle thread runs instead of an application
thread if a debugger related thread blocks. I would have liked to use
priority level 0 for the application threads which would have placed
them below the idle thread but the osThreadSetPriority() API returns
an error for this.

I now run the mriMain() thread at osPriorityISR while it is suspending
and resuming application threads as it is the highest priority level
accepted by the osThreadSetPriority().

Always resume all threads when leaving debug halt mode, even when
single stepping. Previously it only thawed the single stepping thread
but that didn't match the GDB remote stub specification for stop
mode debugging.