Dynamically allocate thread arrays in constructor
Previously the ThreadDebug class allocated a fixed amount of storage to
track the list of active threads. This commit has the ThreadDebug
constructor now take an optional parameter, indicating
the maximum number of threads a developer expects to handle. It
defaults to 32.

I had thought about automatically growing these arrays as needed but
having a debug monitor dynamically allocating memory in a crash
scenario might just make a bad situation worse.

I also thought about just truncating the list if it was too small but
that introduced a bunch of special case handling in the code to deal
with scenarios such as:
* Important application threads not being suspended while halted in
  the debugger.
* GDB not finding the halted thread in the truncated list of threads.
* etc.
I just left in the ASSERT which dumps a message to Serial and then
enters an infinite loop if the arrays are too small to track all of
the active threads.