Add mriIdle() thread
When the debug monitor is active, this thread will take the place of
RTX's idle thread which will be suspended like other application
threads. The mriIdle thread will be suspended again before the
application's threads are resumed so it is only used in place of
RTX's idle thread when halted in GDB.

This change was made so that RTX's idle thread can be treated like any
other application thread which means that GDB is allowed to break into
it if nothing else is running. Without this change, it wouldn't make
sense to break into the code when nothing was actively running and
the idle thread was active since it was a thread that ThreadDebug
wasn't allowed to suspend.