Use major version refs of sketch compilation actions
Previously, due to the lack of a release, the development versions of the sketch compilation actions were used. Using
release versions provides a more stable CI system for the ArduinoCore-mbed project.

Use of the major version ref will cause the workflow to benefit from ongoing development to the action up until such time
as a new major release is made, at which time we would need to evaluate whether any changes to the workflow are required
by the breaking change that triggered the major release before updating the major ref
(e.g., uses: arduino/compile-sketches@v2).