Expand "smoke test" CI workflow to cover new examples
Previously, some of the bundled libraries did not provide example sketches, meaning there was no way to run a
"smoke test" on them via the sketch compilation CI workflow. Now that examples have been added for these libraries, it is
possible to get a basic check that the code compiles.