SocketWrapper: call close() before mis-reusing connection
Some libraries don't call close() and expect the device will do it on its own when the server closes the connection.
To fix this, the best thing to do would be either to use connectionKeepAlive() or to simply close() the client before conneting again.

For backwards compatibility, let's start from scratch every time connect() is colled twice.