Add zoom, flip, and mirror functionality
Adds the functions zoomTo(), zoomToCenter(), setVerticalFlip(),
setHorizontalMirror(), getResolutionWidth(), and
getResolutionHeight() to the Camera library. The old setResolution()
function is changed into the new private function
setResolutionWithZoom(), which is called by the new version of
setResolution(), as well as by the zoomTo() and zoomToCenter()
functions.

These changes required some changes in the abstract class
ImageSensor, which, in turn, required further changes in the camera
drivers. The only camera driver that fully supports the new
functionality is GC2145, and the others return -1 on all calls. The function
signature for setResolution() is changed, and two new functions are
added: setVerticalFlip() and setHorizontalMirror().