Prevent aliasing bugs in the Callback code
This change prevents the GCC compiler from applying optimizations that assume the code follows strict aliasing rules. In order to prevent bugs arising from undefined behavior that is tricky to find in the Callback implementation, or simply from compiler bugs in GCC. The reason for making the change now is concrete customer cases where incorrect machine code was generated in relation to Mbed Callbacks.