Configure post-install script to only operate when platform installed on Linux machine
The Arduino Boards Manager automatically executes the `post_install.sh` script during installation of the platform on a
non-Windows machine.

The platform's post-install script is Linux-specific, but no provisions were made in the script code for the fact that
it is also executed on macOS machines. Previously, this was fairly innocuous because, although misleading, the message
printed to the output during the platform installation was short and cryptic and thus easy for the average macOS user to
ignore:

```
Configuring platform.
Please run as root
```

Since it is important for Linux users to manually run the script and the previous message did not effectively
communicate that, the script was recently modified to print helpful instructions during the Boards Manager installation.
The fact that the script is also executed on macOS machines was not considered in that work. This meant that, although
an improvement for the Linux user experience was accomplished, the macOS user experience was worsened because those
users were then presented with more prominent and detailed inappropriate instructions than the former easy to miss
"Please run as root" instruction. For example

```
Configuring platform.

You might need to configure permissions for uploading.
To do so, run the following command from the terminal:
sudo "/Users/per/Library/Arduino15/packages/arduino/hardware/mbed_giga/4.1.5/post_install.sh"
```

The problem is fixed by adjusting the script so that the script simply returns silently if it is invoked on a non-Linux
machine. macOS users will now only see the following benign message in the output during the Boards Manager
installation:

```
Configuring platform.
```

The POSIX-compliant shell code for determining the operating system the script is running under was derived from the
Arduino CLI application's cross-platform installation script, which has withstood the test of time after years of use by
a large user base.