variants/GENERIC_STM32H747_M4: Update linker_script to support DRAM firmware.
To load the M4 firmware into SDRAM, we must ensure the linker script does not
include any loadable sections that are placed in the shared memory region. If
such sections exist, the ELF loader will overwrite the shared memory region
when loading the firmware.

Additionally, the Open-AMP section is no longer required, as the resource table
is now populated manually in resource_table_init. Therefore, this section can
be removed.

Signed-off-by: iabdalkader <i.abdalkader@gmail.com>