openamp: Load the resource table manually in resource_table_init.
This patch allows removing the Open-AMP section used for the resource table
by loading it manually. The Open-AMP ELF loader would attempt to load this
section if the firmware is built for SDRAM.
Note that since we're not using copy tables or ARM CC, the resource table
has always been copied manually.

Signed-off-by: iabdalkader <i.abdalkader@gmail.com>